+++ shopping-cart-service/src/main/java/n11bootcamp_project_backend/shopping_cart_service/dto/request/UpdateCartItemRequest.java
@@ -1,0 +1,11 @@
+package n11bootcamp_project_backend.shopping_cart_service.dto.request;
+
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+
+public record UpdateCartItemRequest(
+
+        @NotNull(message = "Quantity cannot be null")
+        @Min(value = 1, message = "Quantity must be at least 1")
+        Integer quantity
+) {}