+++ shopping-cart-service/src/main/java/n11bootcamp_project_backend/shopping_cart_service/dto/response/CartItemResponse.java
@@ -1,0 +1,12 @@
+package n11bootcamp_project_backend.shopping_cart_service.dto.response;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public record CartItemResponse(
+        UUID productId,
+        String productName,
+        BigDecimal price,
+        Integer quantity,
+        BigDecimal subTotal
+) {}