+++ shopping-cart-service/src/main/java/n11bootcamp_project_backend/shopping_cart_service/dto/response/CartResponse.java
@@ -1,0 +1,11 @@
+package n11bootcamp_project_backend.shopping_cart_service.dto.response;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.UUID;
+
+public record CartResponse(
+        UUID userId,
+        List<CartItemResponse> items,
+        BigDecimal totalPrice
+) {}