+++ shopping-cart-service/src/test/java/n11bootcamp_project_backend/shopping_cart_service/ShoppingCartServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package n11bootcamp_project_backend.shopping_cart_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ShoppingCartServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}